Rename aggrid.py file
