+++ vizro-core/src/vizro/models/_components/__init__.py
@@ -1,6 +1,6 @@
 """Components that are placed according to the `Layout` of the `Page`."""
 
-from vizro.models._components.aggrid import AgGrid
+from vizro.models._components.ag_grid import AgGrid
 from vizro.models._components.button import Button
 from vizro.models._components.card import Card
 from vizro.models._components.container import Container
